+++ lib/hexapdf/tokenizer.rb
@@ -196,11 +196,13 @@ module HexaPDF
     # Reads the cross-reference subsection entry at the current position and advances the scan
     # pointer.
     #
+    # If a possible problem is detected, yields to caller.
+    #
     # See: PDF1.7 7.5.4
-    def next_xref_entry
+    def next_xref_entry #:yield: matched_size
       prepare_string_scanner(20)
-      unless @ss.skip(/(\d{10}) (\d{5}) ([nf])(?: \r| \n|\r\n)/)
-        raise HexaPDF::MalformedPDFError.new("Invalid cross-reference subsection entry", pos: pos)
+      unless @ss.skip(/(\d{10}) (\d{5}) ([nf])(?: \r| \n|\r\n|\r|\n)/) && @ss.matched_size == 20
+        yield(@ss.matched_size)
       end
       [@ss[1].to_i, @ss[2].to_i, @ss[3]]
     end