+++ test/hexapdf/common_tokenizer_tests.rb
@@ -206,6 +206,8 @@ module CommonTokenizerTests
 
   it "next_xref_entry: fails on invalidly formatted entries" do
     create_tokenizer("0000000001 00001 g \n")
-    assert_raises(HexaPDF::MalformedPDFError) { @tokenizer.next_xref_entry }
+    assert_raises(RuntimeError) { @tokenizer.next_xref_entry { raise }}
+    create_tokenizer("0000000001 00001 n\n")
+    assert_raises(RuntimeError) { @tokenizer.next_xref_entry { raise }}
   end
 end